PR review feedbacks, refactoring
