+++ core/src/main/resources/reference.conf
@@ -450,7 +450,7 @@ jdbc-read-journal {
     query-delay = 1 second
     # The maximum backoff time before trying to query again in case of database failures
     max-backoff-query-delay = 1 minute
-    # The ask timeout to use when querying the journal sequence actor, the actor should normally repond very quickly,
+    # The ask timeout to use when querying the journal sequence actor, the actor should normally respond very quickly,
     # since it always replies with its current internal state
     ask-timeout = 1 second
   }
@@ -550,11 +550,6 @@ jdbc-durable-state-store {
   batchSize = 500
   # New states are retrieved (polled) with this interval.
   refreshInterval = "1s"
-  # stop the live stream query for changes after this number of iterations
-  # to be used only for tests
-  # a value of 0 means don't stop
-  stopAfterEmptyFetchIterations = 0
-
   # can be H2, POSTGRES
   # schemaType = H2
   schemaType = POSTGRES
@@ -589,7 +584,7 @@ jdbc-durable-state-store {
     query-delay = 1 second
     # The maximum backoff time before trying to query again in case of database failures
     max-backoff-query-delay = 1 minute
-    # The ask timeout to use when querying the durable-state sequence actor, the actor should normally repond very quickly,
+    # The ask timeout to use when querying the durable-state sequence actor, the actor should normally respond very quickly,
     # since it always replies with its current internal state
     ask-timeout = 1 second
   }
