+++ core/src/test/scala/akka/persistence/jdbc/state/scaladsl/DataGenerationHelper.scala
@@ -2,7 +2,6 @@ package akka.persistence.jdbc.state.scaladsl
 
 import org.scalatest.concurrent.ScalaFutures
 import scala.concurrent.{ ExecutionContext, Future }
-import akka.dispatch.ExecutionContexts
 
 trait DataGenerationHelper extends ScalaFutures {
 
@@ -47,9 +46,9 @@ trait DataGenerationHelper extends ScalaFutures {
 
     for {
       _ <- Future.unit
-      f1 = Future(upsertManyFor(store, pids.head, tag, 1, noOfItems))(ExecutionContexts.parasitic)
-      f2 = Future(upsertManyFor(store, pids.tail.head, tag, 1, noOfItems))(ExecutionContexts.parasitic)
-      f3 = Future(upsertManyFor(store, pids.last, tag, 1, noOfItems))(ExecutionContexts.parasitic)
+      f1 = Future(upsertManyFor(store, pids.head, tag, 1, noOfItems))
+      f2 = Future(upsertManyFor(store, pids.tail.head, tag, 1, noOfItems))
+      f3 = Future(upsertManyFor(store, pids.last, tag, 1, noOfItems))
       _ <- f1
       _ <- f2
       _ <- f3
